requirements.txt
requests
beautifulsoup4
aiohttp
python-nmap
